pyproject.toml
[project]
name = "benchman"
version = "0.1.0"
description = "Benchmark any AI model — any provider, one command. Terminal + HTML + agent-friendly."
requires-python = ">=3.11"
readme = "README.md"
license = "MIT"
keywords = ["llm", "benchmark", "evaluation", "ai", "cli", "claude", "openai", "ollama"]
authors = [{ name = "Bryan Zane" }]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Environment :: Console",
]
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "questionary>=2.0.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pyarrow>=15.0.0",
]

[project.optional-dependencies]
quality = ["inspect-ai>=0.3.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.5.0"]

[project.scripts]
benchman = "benchman.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/benchman"]

[tool.hatch.build.targets.wheel.force-include]
"src/benchman/data" = "benchman/data"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
